reafctor(updater) add context and improve updater window
Replaces useUpdater context with a new UpdaterContext that provides detailed updater state and types. Updates the splash update page to use the new context and provider props. Sets updater window to always be on top in the main process.
