+++ main/background.ts
@@ -38,6 +38,7 @@ void (async () => {
   const defaultProfile = appSettingsStore.get("defaultProfileUUID");
 
   const updaterWindow = createWindow("updater", {
+    alwaysOnTop: true,
     resizable: !IS_APP_RUNNING_IN_PRODUCTION_MODE ? false : true,
     center: true,
     closable: !IS_APP_RUNNING_IN_PRODUCTION_MODE ? false : true,
